backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.4
django-allauth==0.61.1
django-extensions==3.2.3
django-filter==24.2
django-split-settings==1.3.1
djangorestframework==3.15.1
idna==3.7
Markdown==3.6
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyKakao==0.0.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
